Extract a base class for Seeding integration tests
Extract common setup into `SeedTestBase`, rename existing tests to `SeedScaffoldingTest` to be clear they test the common elements only.
